# Reduce number of queries to fetch `directions`, `user_attrs` and `system_attrs` of study summaries.