@toshihikoyanase 2021-11-22
@ytsmiling @keisuke-umezawa I think you implemented the current code. What do you think of this approach?